+++ src/file.rs
@@ -1,4 +1,4 @@
-mod json;
+mod s3_json;
 mod s3;
 mod utils;
 
@@ -69,7 +69,7 @@ async fn upload_alarms(
         serde_json::to_string(&body.alarms).map_err(|_| FileServerError::SerializationError)?;
     let file_name = format!("alarms_{}_{}.json", id, utils::get_epoch_ms());
     let new_path = format!("{}/{}", constants::ALARM_TABLE_NAME.to_string(), &file_name);
-    let _ = json::save_json(&client, &bucket_name, &new_path, json).await;
+    let _ = s3_json::save_json(&client, &bucket_name, &new_path, json).await;
 
     let old_path = database::alarms::get_alarm_file_path(&connection, id.to_string())
         .await
@@ -90,7 +90,7 @@ async fn upload_alarms(
             .map_err(|err| FileServerError::PostgresDBError {
                 message: err.to_string(),
             })?;
-        let _ = json::delete_json(&client, &bucket_name, &old_path.to_string()).await;
+        let _ = s3_json::delete_json(&client, &bucket_name, &old_path.to_string()).await;
     }
 
     Ok(HttpResponse::Ok().json(StringResponse {
@@ -141,7 +141,7 @@ async fn download_alarms(
 
     log::info!("Download alarms - ID: {}", id);
 
-    let json_str = json::read_json(&client, &bucket_name, &file_path).await?;
+    let json_str = s3_json::read_json(&client, &bucket_name, &file_path).await?;
     let alarms: Vec<Alarm> = serde_json::from_str(&json_str)
         .map_err(|_err| FileServerError::DeserializationError { json_str })?;
 
@@ -199,7 +199,7 @@ async fn delete_alarms(
             message: err.to_string(),
         })?;
 
-    let _ = json::delete_json(&client, &bucket_name, &file_path.to_string()).await;
+    let _ = s3_json::delete_json(&client, &bucket_name, &file_path.to_string()).await;
 
     Ok(HttpResponse::Ok().json(BaseResponse {
         status: 200,
